Adding workaround for gnu core utils treating cp -n as an error ref. #347
Changes to core utils had the side effect of treating the no-clobber option
as an error if the file was not copied if it already exists, thus causing
make to error as well.

Adding this workaround until that issue is addressed.

Alternative solutions:
   - always overwrite the file using the -f / --force option
   - prefix the cp command with a hyphen which will cause make to ignore
     the error, but still report it
   - never copy dwm.desktop during the installation process

+++ Makefile
@@ -66,7 +66,7 @@ endif
 	sed "s/VERSION/${VERSION}/g" < dwm.1 > ${DESTDIR}${MANPREFIX}/man1/dwm.1
 	chmod 644 ${DESTDIR}${MANPREFIX}/man1/dwm.1
 	mkdir -p ${DESTDIR}${PREFIX}/share/xsessions
-	cp -n dwm.desktop ${DESTDIR}${PREFIX}/share/xsessions
+	test -f ${DESTDIR}${PREFIX}/share/xsessions/dwm.desktop || cp -n dwm.desktop ${DESTDIR}${PREFIX}/share/xsessions
 	chmod 644 ${DESTDIR}${PREFIX}/share/xsessions/dwm.desktop
 
 uninstall:
